Removing --become option
I had tried adding a `--become` option to write the file as root on the remote system but I had a bit of trouble getting it working so I mostly reverted those changes for now.  Maybe if I find a good use case and I get inspired, I'll return to it.

I was trying to revert while using my laptop and a remote Linux box I was using for testing - I was editing on both because I can!  I got a little mixed up with the staged commits and had to rebase a couple of times but I think it's all good now.