Improvements - see #108
I noticed that the script didn't like the default date format that Windoze starts with in the USA (MM/DD/YYYY).  I think it expected YYYY-MM-DD because that's what I prefer.  These changes work with the default format.

I added an option to "fake out" the script so it would use some test text rather than running the Windoze `systeminfo` command - that way, I could make the changes on Linux and be pretty sure it would work on Windoze since I put the default output in #108 and knew what it looked like.

I also changed the logging level of a message... no biggie.