Small fix
I was workng up an example of using the class by another Python script and learned that it makes use of my Table class for the --test option.  The Table class is not necessary for use so I changed the explict import into an import that's only used when the --test option is used.