Fixing .ssh/config parsing
The first time I tried the new script at work, it was unable to parse out my ~/.ssh/config.  I only had an AWS instance but the hostname was something like aws-rhel8 and the regular expression was using '\w+' for the hostname which wouldn't accept the hyphen.  I changed the regular expression to '\S+' for the hostname.