Trying to enable --notepad option from WSL
I want to get the script working for Windoze notepad from WSL but I think it's a little harder than I originally thought because the WSL filespace is quite different from Windoze.  Accessing Windoze files from WSL is easy but I'm not sure if there's a way to access a WSL file from Windoze.  This is kind of a work in progress.