Fixing bug
Running ps-tree on WSL, I was sometimes getting strange errors like:

  $ ps-tree $$
  Traceback (most recent call last):
    File "/home/jpfuntne/bin/ps-tree", line 30, in <module>
      processes = ps.Ps().get_processes()
    File "/mnt/c/cygwin64/home/jpfuntne/repos/toys/bin/ps.py", line 104, in get_processes
      'flags': '0x{:08x}'.format(int(stat_tokens[8])),
  TypeError: int() argument must be a string, a bytes-like object or a real number, not 'NoneType'

I think this might have been for zombies or something.  I tried to address this by extending the int_or_none() method to deferring indexing the list until doing the indexing INSIDE the method rather than outside.  But as I write this, I noticed this is one of the few uses of stat_tokens when the int_or_none() method isn't used.  However, I did see it throwing errors yesterday.

So there might still be a bug realted to this specific line and I'll keep and eye on it.