Fixing minor bug
I had made a change a while ago that appended ~/.ssh/config.keep to ~/.ssh/config but injected a bug: the contents of the extra file were added for each and every cloud instance added to ~/.ssh/config.  I'm changing the logic to append **after** all of the cloud instances have been added.