Turning aws assert into a warning message
After I switched to a method that getting AWS credentials that are only good for 8 hours (!!), I added an `assert` to let me know when the aws cli fails.  Instead of failing, I'm just going to turn the `assert` into a warning message and think I will lean on GCP instances more moving forward.  This won't require me to renew the credentials so often.