Switching over to using find shell command
Originally the script would search for the pattern mostly using Python code.  After I switched my work machine to WSL, this proved to be too slow but using the shell find command works better.  The command only uses glob-style patterns as opposed to regular expressions but I think that's a fine compromise.

At first I was thinking I would make use of the -follow predicate of the find command but I didn't think it was necessary.  I can add it if I want.