Tweak do_help() for do_EOF()
I didn't add doc to do_EOF() and didn't really plan to but this caused a problem for do_help() because it would spin through the commands and look for the doc.  Since do_EOF() didn't have doc, do_help() would below up when it tried to process it.  I added a condition to skip commands that don't have doc.